requirements.txt
arxiv>=2.1.0
google-genai>=1.0.0
pyyaml>=6.0
